blossom_socket/requirements.txt
click==8.1.3
Django==4.0.5
flask==2.1.2
Flask-Cors==3.0.10
Flask-SocketIO==5.2.0
gcloud==0.18.3
getch==1.0
gunicorn==20.1.0
Jinja2==3.1.2
prettytable==3.3.0
py-getch==1.0.1
pypot==5.0.2
six==1.16.0
sseclient==0.0.24
Werkzeug==2.1.2
python-jwt==3.3.2
pycrypto==2.6.1
PyYAML==6.0
python-engineio==4.3.2
python-socketio==5.6.0
itsdangerous==2.1.2
